requirements.txt
# Core Django and Web Framework
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
channels>=4.0.0
channels-redis>=4.1.0

# Database and Caching
psycopg2-binary>=2.9.7
redis>=5.0.0

# Authentication and Security
djangorestframework-simplejwt>=5.3.0
cryptography>=41.0.0
django-ratelimit>=4.1.0
django-oauth-toolkit>=1.7.0

# Media Processing and Storage
Pillow>=10.0.0
boto3>=1.28.0
django-storages>=1.14.0

# Real-time Communication
websockets>=11.0.0
aioredis>=2.0.0

# API Documentation
drf-spectacular>=0.26.0

# Development and Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Monitoring and Logging
sentry-sdk>=1.32.0
django-extensions>=3.2.0

# Environment Management
python-decouple>=3.8

# Async Support
asgiref>=3.7.0
uvicorn>=0.23.0

# Image Processing and AI
opencv-python>=4.8.0
numpy>=1.24.0

# FFmpeg and Media Processing
ffmpeg-python>=0.2.0
pydub>=0.25.1
moviepy>=1.0.3
Pillow>=10.0.0

# Celery for Background Tasks
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Additional Media Libraries
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# AI and Machine Learning
openai>=1.0.0
google-cloud-vision>=3.4.0
boto3>=1.28.0
botocore>=1.31.0

# Security and Encryption
cryptography>=41.0.0
pyjwt>=2.8.0

# Mobile and Push Notifications
firebase-admin>=6.2.0
aiohttp>=3.8.0

# Analytics and Monitoring
psutil>=5.9.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Cloud Storage
boto3>=1.28.0
botocore>=1.31.0
